requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
seaborn==0.13.2
streamlit==1.40.2
together==1.3.5
python-dotenv==1.0.1
matplotlib==3.9.2
chardet==5.2.0
openai==1.54.1
anthropic==0.49.0
google-generativeai==0.8.4
regex==2024.9.11
httpx==0.27.2
xlsxwriter==3.2.0
transformers==4.48.3
vllm==0.7.3
torch==2.5.1
torchvision==0.20.1
sentencepiece==0.2.0
sacremoses==0.1.1
openpyxl==3.1.5
krippendorff==0.8.1
google-genai==1.50.1